// TTT_WebFrameWorks/src/types.ts
export interface Duckie {
    id: string;
    name: string;
    type: string;
    description: string;
    price: number;
    colorPalette: string[];
    materialBlend: string;
    accessories: string[];
    mood: string;
    waterBehavior: string;
    soundProfile: string;
    signatureMove: string;
    image: string;
}